add function documentation and game turn metadata
