+++ src-tauri/src/chess/mod.rs
@@ -12,10 +12,11 @@ pub mod data {
     // pub struct History(Vec<i32>);
 
     /// game score stored relative to white
-    pub struct Score(pub Mutex<i32>);
 
     /// game state 8x8 board
     pub struct PieceLocation(pub Mutex<super::types::BoardState>);
 
     pub struct SelectedSquare(pub Mutex<Option<super::types::Square>>);
+
+    pub struct GameMeta(pub Mutex<super::types::GameMeta>);
 }
