+++ src-tauri/src/chess/types.rs
@@ -4,6 +4,17 @@ use serde::{Deserialize, Serialize};
 
 pub type BoardState = [[Piece; 8]; 8];
 
+pub struct GameMeta {
+    pub turn: usize,
+    pub score: i32,
+}
+
+impl Default for GameMeta {
+    fn default() -> Self {
+        GameMeta { turn: 0, score: 0 }
+    }
+}
+
 /// Square reference in row and column
 pub type Square = (usize, usize);
 pub type IsAttack = bool;
