+++ src-tauri/src/main.rs
@@ -8,7 +8,7 @@ mod chess;
 fn main() {
     tauri::Builder::default()
         .manage(chess::data::PieceLocation(Default::default()))
-        .manage(chess::data::Score(Default::default()))
+        .manage(chess::data::GameMeta(Default::default()))
         .manage(chess::data::SelectedSquare(Default::default()))
         .invoke_handler(tauri::generate_handler![
             chess::board::new_game,
